requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
ecdsa==0.19.0
fastapi==0.101.0
h11==0.14.0
idna==3.7
passlib==1.7.4
pyasn1==0.6.0
pydantic==1.10.4
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.46
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.23.2
